requirements.txt
darts==0.29.0
pandas==2.2.2
python_dateutil==2.9.0.post0
pytorch_lightning==2.2.3
